pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sam-clip-rag"
version = "0.1.0"
description = "Multimodal bamboo species recognition with SAM, CLIP, and Milvus"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "numpy>=1.26.0",
  "Pillow>=10.0.0",
  "torch>=2.2.0",
  "transformers>=4.44.0",
  "pymilvus>=2.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.setuptools.packages.find]
where = ["src"]
